third_party/blink/web_tests/external/wpt/editing/run/forwarddelete_1-1000-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["forwarddelete",""]] "<span>foo[</span><span>]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span>foo</span><span>bar</span>" but got "<span>foo</span><span>ar</span>"
[FAIL] [["stylewithcss","true"],["forwarddelete",""]] "foo[]<span style=display:none>bar</span>baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["defaultparagraphseparator","div"],["forwarddelete",""]] "foo{}<p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo" but got "foo<p></p>"
[FAIL] [["defaultparagraphseparator","p"],["forwarddelete",""]] "foo{}<p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo" but got "foo<p></p>"
Harness: the test ran to completion.
